# Rework optimizer parameters