… fvdb_reality_capture/training/sfm_dataset.py
… fvdb_reality_capture/transforms/__init__.py
… fvdb_reality_capture/transforms/crop_scene.py
… fvdb_reality_capture/transforms/downsample_images.py
… tests/unit/test_gaussian_splat_optimizer.py
… train.py
… visualize.py