**The conclusion:**
I think we should keep the default settings as they are today but expose these new parameters for users who want them. The new defaults are not always better but they seem somewhat consistent and I don't want to rock the boat too much before release. 

**Genesis of these changes**
The optimizer used to use fixed threshold for gradient pruning. This turns out to be sensitive to the image resolution. This PR sets the threshold based on a percentile of the gradient distribution (this can be fixed once or updated on every refinement step). It also changes pose optimization to not start until refinement is done (which kind of makes more sense). 

## Motivation and description of these changes.

**Background** The `GausssianSplatOptimizer` performs a refinement step every so many epochs. In this step, certain Gaussians are either (1) _duplicated_ into two exact copies of each other (2) _subdivided_ into two smaller Gaussians, or (3) _pruned_ entirely from the scene. For each Gaussian during refinement we use a different heuristic to decide if it should be duplicated, subdivided, or pruned. _This change has to do with duplication_ and _subdivision_ but I'll discuss the heuristics for all three actions for completion.
 1. We duplicate a Gaussian if it's 3D size is below some threshold and the gradient of its projected means over time is high on average. Intuitively, this means the Gaussian is not taking up a lot of space in the scene, but consistently wants to change positions when viewed from different images. Likely this Gaussian is stuck trying to represent too much of the scene and should be split into two copies.
 2. We subdivide a Gaussian when its 3D size exceeds a threshold value and the gradient of its projected mean over time is high on average. In this case, a Gaussian is likely too large for the amount of detail it represents and should be split to capture detail in the image.
 3. We prune a Gaussian if its opacity falls below a threshold since it is not contributing much to the image.

For actions (1) and (2), we use the average gradient of projected means of the Gaussians over time to decide if a Gaussian should be pruned or not. To do this, whenever we compute gradients, we store the gradients of the 2d projection of the means and accumulate them over time. This value is then thresholded to decide whether to refine or not. 

**The status quo**
The original optimizer uses a default fixed threshold of 0.002 to determine if an average gradient is high. We found this value often leads to the optimizer capping out the number of Gaussians at a relatively small value, which causes reconstructions to not capture detail in a scene. 

After this change, we offer three modes for setting this threshold: 
 1. Using a fixed threshold as before.
 2. Using the N^th percentile of the distribution of 2d mean gradients at every refinement step (thus adaptively refining to capture more and more detail as the scene trains. This produces many more Gaussians and reconstructions that capture much more detail.
 3. Using the N^th percentile of the distribution of 2d mean gradients at the first refinement step as a fixed threshold (which exhibits similar behavior as the status quo for many scenes but tends to be a more robust choice)

I ran some initial experiments on the 360v2 dataset which consists mostly of small object centric captures.
In this case, option (3) is slightly faster with marginally better quality which also being more principled. Option (1) is comparable (likely because it was tuned on this data).

Option (2) performs poorly (slow and lower PSNR) but I think it's because I'm not increasing the number of training iterations to compensate for the greater number of Gaussians. We have evidence showing option (2) works well if you run training for longer, since option (2) allows for very large numbers of Gaussians, these also require more training time to converge.

## Benchmark results
Option 1: Old default
Option 2: Fully adaptive refinement
Option 3: Automatic static parameter selection based on gradient histogram

#### Limitations:
We did not run these benchmarks until refinement plateaus which means some of these (in particular option 2) may be leaving money on the table. 
This doesn't demonstrate that Option 2 can produce many more gaussians which is needed in some cases.
We should probably only turn on pose optimization when we stop refining, which improves quality a lot when gaussian growth happens at a larger rate since learning is noiser.

#### TL;DR
On 360v2, Option 3 produces generally better quality results with fewere Gaussians.
On Gettysburg, the Option 1 is faster, produce fewer gaussians, and is higher quality. 
On Safety Park, Option 1 is a bit faster than Option 3 with similar quality. 


**Benchmark on 360v2**
```
Training time (s):
 - Method 1 (old default): Mean 800.02s, Median 644.99s
 - Method 2:  Mean 1924.1s, Median 1996.3s
 - Method 3: Mean 751.02s, Median 601.17s

PSNR (dB):
 - Method 1: (old default) Mean 28.52dB, Median 28.73dB
 - Method 2: Mean: 24.7dB, Median 27.1dB
 - Method 3: Mean 28.56dB, Median 28.78dB

SSIM (%):
 - Method 1 (old default): Mean 86.8%, Median 90.2%
 - Method 2: Mean 70%, Median 90%
 - Method 3: 86.7%, Median 90.5%

Number of Gaussians:
 - Method 1 (old default): Mean 2,652,434, Median 2,652,434
 - Method 2: Mean 7,003,206, Median 7,194,541
 - Method 3: Mean 2,368,494, Median 1,325,510
```
Plot for this benchmark (I ran the adaptive run separately so it's not on this graph)
<img width="2148" height="9152" alt="image" src="https://github.com/user-attachments/assets/0f85f211-6cbe-4284-85a3-8b925900d63c" />


**Benchmark on gettysburg and safety_park**
```
Training time (s):
 - Method 1 (old default): Mean 1235.3s, Median 1235.3s
 - Method 2:  Mean 2177.6s, Median 2177.6s
 - Method 3: Mean 1817.8s, Median 1817.8s

PSNR (dB):
 - Method 1: (old default) Mean 30.6dB, Median 30.6dB
 - Method 2: Mean: 30.9dB, Median 30.9dB
 - Method 3: Mean 29.3dB, Median 29.3dB

SSIM (%):
 - Method 1 (old default): Mean 90%, Median 90%
 - Method 2: Mean 90%, Median 90%
 - Method 3: 90%, Median 90%

Number of Gaussians:
 - Method 1 (old default): Mean 3,423,000, Median 3,423,000
 - Method 2: Mean 9,311,195, Median 9,311,195
 - Method 3: Mean 5,783,831, Median 5,783,831
```
Benchmark plot: 
<img width="2148" height="9152" alt="image" src="https://github.com/user-attachments/assets/b4a8b263-aed5-4be1-8595-c6fe9e30c07a" />

This suggests that we may want to settle on option (3) as a default. Before doing this, I want to run the benchmark on gettysburg safety_park, and miris_factory which are outdoor drone capture and large scale indoor capture.